refactor: Remove legacy process instance component
